Added support for older versions of Discourse
I initially implemented the upload functionality by copying the code
in https://github.com/discourse/discourse-insert-video. The will
currently only work for the upcoming 3.4.0-beta branch however.

To support more Discourse versions this changeset therefore switches
to an older implementation of the upload code. We can revert to the
modern way of doing it when it has wider support.